refactor: update shader instance ID types to use unsigned integers
- Changed instance ID types in EntityUIDOutputSingleShader.frag and common shader files from float to uint for consistency with recent updates.
- Ensured compatibility across shader functions by updating parameter types for instance IDs, enhancing performance and clarity in shader logic.

+++ src/webgpu/shaderity_shaders/common/iblDefinition.wgsl
@@ -142,7 +142,7 @@ fn get_sample_from_backbuffer(sampleCoord: vec2f, perceptualRoughness: f32, ior:
   return transmittedLight;
 }
 
-fn getIBLVolumeRefraction(baseColor: vec3f, normal: vec3f, view: vec3f, cameraSID: u32, materialSID: u32, thickness: f32, perceptualRoughness: f32, ior: f32, attenuationColor: vec3f, attenuationDistance: f32, position_inWorld: vec3f, instanceIds: vec4<f32>) -> vec3f {
+fn getIBLVolumeRefraction(baseColor: vec3f, normal: vec3f, view: vec3f, cameraSID: u32, materialSID: u32, thickness: f32, perceptualRoughness: f32, ior: f32, attenuationColor: vec3f, attenuationDistance: f32, position_inWorld: vec3f, instanceIds: vec4<u32>) -> vec3f {
 #ifdef RN_USE_DISPERSION
   let dispersion = get_dispersion(materialSID, 0);
   let halfSpread = (ior - 1.0) * 0.025 * dispersion;
@@ -217,7 +217,7 @@ fn getIBLFresnelGGX(perceptualRoughness: f32, NdotV: f32, F0: vec3f, specularWei
   return FssEss + FmsEms;
 }
 
-fn IBLContribution(instanceIds: vec4<f32>, materialSID: u32, cameraSID: u32,
+fn IBLContribution(instanceIds: vec4<u32>, materialSID: u32, cameraSID: u32,
   normal_inWorld: vec3f, NdotV: f32, viewDirection: vec3f, geomNormal_inWorld: vec3f, position_inWorld: vec3f,
   baseColor: vec3f, perceptualRoughness: f32, metallic: f32, specularWeight: f32, dielectricF0: vec3f, ior: f32,
   clearcoatProps: ClearcoatProps,
